+++ rally_openstack/environment/platforms/existing.py
@@ -321,7 +321,7 @@ class OpenStack(platform.Platform):
 
         """
 
-        from oslo_utils import strutils
+        from oslo_utils import strutils  # type: ignore[import-untyped]
 
         required_env_vars = ["OS_AUTH_URL", "OS_USERNAME", "OS_PASSWORD"]
         missing_env_vars = [v for v in required_env_vars if
@@ -395,7 +395,7 @@ class OpenStack(platform.Platform):
         env_vars_description = "\n".join(
             line for line in env_vars_docs.split("\n")[1:]
         )
-        doc += (f"\n **The following environment variables are expected for "
+        doc += (f"\n\n**The following environment variables are expected for "
                 f"creation a Rally environment using sustem environment "
                 f"variables**\n{env_vars_description}")
         return doc