+++ rally_openstack/verification/tempest/consts.py
@@ -18,9 +18,7 @@ from rally.common import utils
 
 class _TempestApiTestSets(utils.ImmutableMixin, utils.EnumMixin):
     BAREMETAL = "baremetal"
-    CLUSTERING = "clustering"
     COMPUTE = "compute"
-    DATA_PROCESSING = "data_processing"
     DATABASE = "database"
     IDENTITY = "identity"
     IMAGE = "image"