+++ tests/unit/verification/tempest/test_config.py
@@ -76,18 +76,6 @@ class TempestConfigfileManagerTestCase(test.TestCase):
         for item in expected:
             self.assertIn(item, result)
 
-    @ddt.data("data_processing", "data-processing")
-    def test__configure_data_processing(self, service_type):
-        self.tempest.available_services = ["sahara"]
-
-        self.tempest.clients.services.return_value = {
-            service_type: "sahara"}
-        self.tempest.conf.add_section("data-processing")
-        self.tempest._configure_data_processing()
-        self.assertEqual(service_type,
-                         self.tempest.conf.get("data-processing",
-                                               "catalog_type"))
-
     @ddt.data(
         # The prefix "ex_" is abbreviation of "expected"
         # case #1: both versions are discoverable; version is in the auth_url
@@ -170,6 +158,17 @@ class TempestConfigfileManagerTestCase(test.TestCase):
     def test__configure_network_if_neutron(self):
         self.tempest.available_services = ["neutron"]
         client = self.tempest.clients.neutron()
+        client.list_rbac_policies.return_value = {
+            "rbac_policies": [
+                {
+                    "target_tenant": "*",
+                    "object_type": "network",
+                    "object_id": "test_id",
+                    "action": "access_as_shared",
+                    "id": "test_id"
+                }
+            ]
+        }
         client.list_networks.return_value = {
             "networks": [
                 {
@@ -247,15 +246,14 @@ class TempestConfigfileManagerTestCase(test.TestCase):
             self.assertIn(item, result)
 
     def test__configure_service_available(self):
-        available_services = ("nova", "cinder", "glance", "sahara")
+        available_services = ("nova", "cinder", "glance")
         self.tempest.available_services = available_services
         self.tempest.conf.add_section("service_available")
         self.tempest._configure_service_available()
 
         expected = (
             ("neutron", "False"), ("heat", "False"), ("nova", "True"),
-            ("swift", "False"), ("cinder", "True"), ("sahara", "True"),
-            ("glance", "True"))
+            ("swift", "False"), ("cinder", "True"), ("glance", "True"))
         result = self.tempest.conf.items("service_available")
         for item in expected:
             self.assertIn(item, result)