+++ CHANGELOG.rst
@@ -16,6 +16,38 @@ Changelog
 .. Release notes for existing releases are MUTABLE! If there is something that
    was missed or can be improved, feel free to change it!
 
+[unreleased]
+------------
+
+Added
+~~~~~
+
+* CI jobs for checking compatibility with python 3.12 and 3.13
+
+Removed
+~~~~~~~
+
+* Support for Python 3.8 and Python 3.9 is dropped
+* Removed all support for the retired Murano project
+* Removed all support for the retired Sahara project
+* Removed all support for the retired Senlin project
+* Removed all support for the retired Monasca project
+
+Changed
+~~~~~~~
+
+* Implements pep-517 (pyproject.toml) and replaces pbr dependency
+  with setuptools-scm
+* Zaqar scenarios now use messaging v2 API, instead of deprecated v1 API.
+* Bump minimal required version to Rally 5.0.0. Switch docker image to use it.
+* All task samples are ported to Task format V2
+
+Fixed
+~~~~~
+
+* Fix restoring quotas bug while rally cleanup
+* Don't load compute resources in tempest verifier if nova is not enabled
+
 [3.0.0] - 2024-05-23
 --------------------